+++ stella/src/emucore/CartF4.cxx
@@ -13,7 +13,7 @@
 // See the file "license" for information on usage and redistribution of
 // this file, and for a DISCLAIMER OF ALL WARRANTIES.
 //
-// $Id: CartF4.cxx,v 1.6 2005-10-12 03:32:28 urchlay Exp $
+// $Id: CartF4.cxx,v 1.7 2005-12-17 01:23:07 stephena Exp $
 //============================================================================
 
 #include <assert.h>
@@ -153,7 +153,7 @@ bool CartridgeF4::save(Serializer& out)
   try
   {
     out.putString(cart);
-    out.putLong(myCurrentBank);
+    out.putInt(myCurrentBank);
   }
   catch(char* msg)
   {
@@ -181,7 +181,7 @@ bool CartridgeF4::load(Deserializer& in)
       return false;
     }
 
-    myCurrentBank = (uInt16)in.getLong();
+    myCurrentBank = (uInt16)in.getInt();
   }
   catch(char* msg)
   {
